Fix issues with failing Wikidata logins
Wikidata can sometimes be grumpy and disallow logging-in because of database lag (gib us galera, wmf). To avoid stalling the bot each time that happens, I've moved the login command to a later stage in the code. I've also reduced the max. number of retries, so the bot doesn't stumble over itself when it's restarted by the scheduler.

It should be far less annoying with error messages.